Allow BigInts and Uint8Arrays to be specified as custom context variables. Deserialization of hex strings leave them unchanged. Move to/fromJSON out of shared since only applicable to taco package. Update tests accordingly.
