+++ packages/taco/test/conditions/condition-expr.test.ts
@@ -1,5 +1,5 @@
 import { initialize } from '@nucypher/nucypher-core';
-import { objectEquals, toJSON } from '@nucypher/shared';
+import { objectEquals } from '@nucypher/shared';
 import { USER_ADDRESS_PARAM_DEFAULT } from '@nucypher/taco-auth';
 import { TEST_CHAIN_ID, TEST_CONTRACT_ADDR } from '@nucypher/test-utils';
 import { SemVer } from 'semver';
@@ -19,6 +19,7 @@ import {
 import { CompoundCondition } from '../../src/conditions/compound-condition';
 import { ConditionExpression } from '../../src/conditions/condition-expr';
 import { ERC721Balance } from '../../src/conditions/predefined/erc721';
+import { toJSON } from '../../src/utils';
 import {
   testContractConditionObj,
   testFunctionAbi,
